requirements.txt
python-dotenv
requests
fastmcp